updated test script with better lr
